// packages/wizzi.demo.server/src/features/wizziMeta/types.ts
/*
    artifact generator: C:\My\wizzi\stfnbssl\wizzi\packages\wizzi-js\lib\artifacts\ts\module\gen\main.js
    package: wizzi-js@0.7.14
    primary source IttfDocument: C:\My\wizzi\stfnbssl\wizzi.demo\packages\wizzi.demo.server\.wizzi\src\features\wizziMeta\types.ts.ittf
*/
export type WizziMetaOptions = { 
    metaSourcesFolderPath?: string;
    metaGenerator?: string;
    cliCtx?: any;
    cliCtxName?: string;
    outputPackageName?: string;
    outputPackagePath?: string;
    description?: string;
};